delete : 성능 최적화 코드 제거
