+++ src/app/(pages)/login/page.tsx
@@ -1,16 +1,11 @@
 'use client';
 
-import dynamic from 'next/dynamic';
 import Image from 'next/image';
 
 import { GoogleLoginButton } from './_components/GoogleLoginButton';
 import { KakaoLoginButton } from './_components/KakaoLoginButton';
 import { LoginErrorBanner } from './_components/LoginErrorBanner';
-
-const LoginVisual = dynamic(
-  () => import('./_components/LoginVisual').then((mod) => ({ default: mod.LoginVisual })),
-  { ssr: false },
-);
+import { LoginVisual } from './_components/LoginVisual';
 
 export default function LoginPage() {
   return (
